+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 1.1.3 (2019-01-27)
+
+#### Added
+
+- filter button to show/hide finished Todo tasks from list
+- button to start time log for a specific todo task
+
 ## 1.1.3 (2019-01-22)
 
 #### Fixed